+++ changelog.md
@@ -6,6 +6,8 @@
 This version is not released yet and is under active development.
 ```
 
+- Fix auto-mapping and recognition of `click.IntRange` options in config module. Closes {issue}`170`.
+
 ## {gh}`2.1.1 (2022-05-22) <compare/v2.1.0...v2.1.1>`
 
 - Fix compatibility with `cloup >= 0.14.0`.